--- KL25Z Code/Objects/project.build_log.htm
@@ -1,75 +1,0 @@
-<html>
-<body>
-<pre>
-<h1>µVision Build Log</h1>
-<h2>Tool Versions:</h2>
-IDE-Version: ÂµVision V5.29.0.0
-Copyright (C) 2019 ARM Ltd and ARM Germany GmbH. All rights reserved.
-License Information: Justin Wong, NUS, LIC=----
- 
-Tool Versions:
-Toolchain:       MDK-Lite  Version: 5.29.0.0
-Toolchain Path:  C:\Keil_v5\ARM\ARMCC\Bin
-C Compiler:      Armcc.exe V5.06 update 6 (build 750)
-Assembler:       Armasm.exe V5.06 update 6 (build 750)
-Linker/Locator:  ArmLink.exe V5.06 update 6 (build 750)
-Library Manager: ArmAr.exe V5.06 update 6 (build 750)
-Hex Converter:   FromElf.exe V5.06 update 6 (build 750)
-CPU DLL:         SARMCM3.DLL V5.29.0.0
-Dialog DLL:      DARMCM1.DLL V1.19.2.0
-Target DLL:      CMSIS_AGDI.dll V1.32.0.0
-Dialog DLL:      TARMCM1.DLL V1.14.1.0
- 
-<h2>Project:</h2>
-C:\Users\justi\OneDrive\Documents\NUS\Year 2 Sem 1\CG2271\Project\Code\KL25Z Code\project.uvprojx
-Project File Date:  10/15/2021
-
-<h2>Output:</h2>
-*** Using Compiler 'V5.06 update 6 (build 750)', folder: 'C:\Keil_v5\ARM\ARMCC\Bin'
-Build target 'Target 1'
-compiling project.c...
-constants.h(11): warning:  #1-D: last line of file ends without a newline
-  } directions;
-project.c(41): warning:  #188-D: enumerated type mixed with another type
-  			dir = (serialValue & 0b11101111);
-project.c(217): warning:  #111-D: statement is unreachable
-  	return 1;
-project.c(218): warning:  #1-D: last line of file ends without a newline
-  }
-project.c: 4 warnings, 0 errors
-linking...
-Program Size: Code=2660 RO-data=224 RW-data=8 ZI-data=352  
-".\Objects\project.axf" - 0 Error(s), 4 Warning(s).
-
-<h2>Software Packages used:</h2>
-
-Package Vendor: ARM
-                http://www.keil.com/pack/ARM.CMSIS.5.6.0.pack
-                ARM.CMSIS.5.6.0
-                CMSIS (Cortex Microcontroller Software Interface Standard)
-   * Component: CORE Version: 5.3.0
-
-Package Vendor: Keil
-                http://www.keil.com/pack/Keil.Kinetis_KLxx_DFP.1.15.0.pack
-                Keil.Kinetis_KLxx_DFP.1.15.0
-                NXP Kinetis KLxx Series Device Support and Examples
-   * Component: Startup Version: 2.5.0
-
-<h2>Collection of Component include folders:</h2>
-  .\RTE\Device\MKL25Z128xxx4
-  .\RTE\_Target_1
-  C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include
-  C:\Users\justi\AppData\Local\Arm\Packs\Keil\Kinetis_KLxx_DFP\1.15.0\Device\Include
-
-<h2>Collection of Component Files used:</h2>
-
-   * Component: ARM::CMSIS:CORE:5.3.0
-
-   * Component: Keil::Device:Startup:2.5.0
-      Source file:   Device\Source\ARM\startup_MKL25Z4.s
-      Source file:   Device\Source\system_MKL25Z4.c
-      Include file:  Device\Source\system_MKL25Z4.h
-Build Time Elapsed:  00:00:01
-</pre>
-</body>
-</html>
