--- KL25Z Code/Objects/project.d
@@ -1,9 +1,0 @@
-.\objects\project.o: project.c
-.\objects\project.o: C:\Users\justi\AppData\Local\Arm\Packs\Keil\Kinetis_KLxx_DFP\1.15.0\Device\Include\MKL25Z4.h
-.\objects\project.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
-.\objects\project.o: C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\core_cm0plus.h
-.\objects\project.o: C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_version.h
-.\objects\project.o: C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_compiler.h
-.\objects\project.o: C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_armcc.h
-.\objects\project.o: .\RTE\Device\MKL25Z128xxx4\system_MKL25Z4.h
-.\objects\project.o: constants.h
