--- KL25Z Code/Objects/project.lnp
@@ -1,7 +1,0 @@
---cpu Cortex-M0+
-".\objects\project.o"
-".\objects\startup_mkl25z4.o"
-".\objects\system_mkl25z4.o"
---ro-base 0x00000000 --entry 0x00000000 --rw-base 0x1FFFF000 --entry Reset_Handler --first __Vectors --strict --summary_stderr --info summarysizes --map --load_addr_map_info --xref --callgraph --symbols
---info sizes --info totals --info unused --info veneers
---list ".\Listings\project.map" -o .\Objects\project.axf
