--- KL25Z Code/Objects/startup_mkl25z4.d
@@ -1,1 +1,0 @@
-.\objects\startup_mkl25z4.o: RTE\Device\MKL25Z128xxx4\startup_MKL25Z4.s
