--- KL25Z Code/Objects/system_mkl25z4.d
@@ -1,8 +1,0 @@
-.\objects\system_mkl25z4.o: RTE\Device\MKL25Z128xxx4\system_MKL25Z4.c
-.\objects\system_mkl25z4.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
-.\objects\system_mkl25z4.o: C:\Users\justi\AppData\Local\Arm\Packs\Keil\Kinetis_KLxx_DFP\1.15.0\Device\Include\MKL25Z4.h
-.\objects\system_mkl25z4.o: C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\core_cm0plus.h
-.\objects\system_mkl25z4.o: C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_version.h
-.\objects\system_mkl25z4.o: C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_compiler.h
-.\objects\system_mkl25z4.o: C:\Users\justi\AppData\Local\Arm\Packs\ARM\CMSIS\5.6.0\CMSIS\Core\Include\cmsis_armcc.h
-.\objects\system_mkl25z4.o: .\RTE\Device\MKL25Z128xxx4\system_MKL25Z4.h
