+++ KL25Z Code/PWM.h
@@ -1,0 +1,12 @@
+#ifndef PWM_HEADER
+#define PWM_HEADER
+
+#include "foundation.h"
+
+extern volatile int speed;
+
+void initPWM1(void);
+
+void startPWM(void);
+
+#endif
