+++ KL25Z Code/RTE/Device/MKL25Z128xxx4/system_MKL25Z4.h
@@ -99,13 +99,13 @@
 #ifndef SYSTEM_MKL25Z4_H_
 #define SYSTEM_MKL25Z4_H_                        /**< Symbol preventing repeated inclusion */
 
-#define CLOCK_SETUP 1
-
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 #include <stdint.h>
+
+#define CLOCK_SETUP 1
 
 
 #ifndef DISABLE_WDOG
