+++ KL25Z Code/RTE/_Target_1/RTE_Components.h
@@ -3,7 +3,7 @@
  * Auto generated Run-Time-Environment Configuration File
  *      *** Do not modify ! ***
  *
- * Project: 'project' 
+ * Project: 'RtosProject' 
  * Target:  'Target 1' 
  */
 
@@ -16,6 +16,9 @@
  */
 #define CMSIS_device_header "MKL25Z4.h"
 
+/*  ARM::CMSIS:RTOS2:Keil RTX5:Library:5.5.1 */
+#define RTE_CMSIS_RTOS2                 /* CMSIS-RTOS2 */
+        #define RTE_CMSIS_RTOS2_RTX5            /* CMSIS-RTOS2 Keil RTX5 */
 
 
 #endif /* RTE_COMPONENTS_H */
