+++ KL25Z Code/UART.h
@@ -1,0 +1,17 @@
+#ifndef UART_HEADER
+#define UART_HEADER
+
+#include "foundation.h"
+#include "constants.h"
+
+#define BAUD_RATE 9600
+#define UART_TX_PORTE 22
+#define UART_RX_PORTE 23
+
+extern volatile tasks dir;
+
+void initUART2(uint32_t);
+
+void UART2_IRQHandler(void);
+
+#endif
