+++ KL25Z Code/motors.c
@@ -1,0 +1,33 @@
+#include "motors.h"
+
+void initMotors() {
+	SIM->SCGC5 |= (SIM_SCGC5_PORTA_MASK | SIM_SCGC5_PORTC_MASK | SIM_SCGC5_PORTD_MASK);
+	
+	PORTA->PCR[16] &= ~PORT_PCR_MUX_MASK;
+	PORTA->PCR[16] |= PORT_PCR_MUX(1);
+	
+	PORTA->PCR[17] &= ~PORT_PCR_MUX_MASK;
+	PORTA->PCR[17] |= PORT_PCR_MUX(1);
+	
+	PORTC->PCR[16] &= ~PORT_PCR_MUX_MASK;
+	PORTC->PCR[16] |= PORT_PCR_MUX(1);
+	
+	PORTC->PCR[17] &= ~PORT_PCR_MUX_MASK;
+	PORTC->PCR[17] |= PORT_PCR_MUX(1);
+	
+	PORTD->PCR[0] &= ~PORT_PCR_MUX_MASK;
+	PORTD->PCR[0] |= PORT_PCR_MUX(1);
+	
+	PORTD->PCR[1] &= ~PORT_PCR_MUX_MASK;
+	PORTD->PCR[1] |= PORT_PCR_MUX(1);
+	
+	PORTD->PCR[2] &= ~PORT_PCR_MUX_MASK;
+	PORTD->PCR[2] |= PORT_PCR_MUX(1);
+	
+	PORTD->PCR[3] &= ~PORT_PCR_MUX_MASK;
+	PORTD->PCR[3] |= PORT_PCR_MUX(1);
+	
+	PTA->PDDR |= ((1 << 16) | (1 << 17));
+	PTC->PDDR |= ((1 << 16) | (1 << 17));
+	PTD->PDDR |= (0b1111);
+}
