+++ KL25Z Code/motors.h
@@ -1,0 +1,8 @@
+#ifndef MOTORS_HEADER
+#define MOTORS_HEADER
+
+#include "foundation.h"
+
+void initMotors(void);
+
+#endif
